Refactor gguf_util (#2552) (#2552)
Summary: Pull Request resolved: https://github.com/pytorch/executorch/pull/2552

Differential Revision: D55167953

Pulled By: mergennachin
